Move common option code to shared package
