+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/impl/option/BlackOneTouchAssetPriceFormulaRepository.java
@@ -11,7 +11,7 @@ import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.math.impl.statistics.distribution.NormalDistribution;
 import com.opengamma.strata.math.impl.statistics.distribution.ProbabilityDistribution;
-import com.opengamma.strata.product.fxopt.SimpleConstantContinuousBarrier;
+import com.opengamma.strata.product.option.SimpleConstantContinuousBarrier;
 
 /**
  * The price function to compute the price of one-touch or no-touch (asset-or-nothing) option in the Black world.
