+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/impl/tree/ConstantContinuousSingleBarrierKnockoutFunction.java
@@ -26,7 +26,7 @@ import org.joda.beans.impl.direct.DirectMetaPropertyMap;
 import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.collect.array.DoubleArray;
 import com.opengamma.strata.product.common.PutCall;
-import com.opengamma.strata.product.fxopt.BarrierType;
+import com.opengamma.strata.product.option.BarrierType;
 
 /**
  * Single barrier knock-out option function.
