+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/index/IborFutureOptionMarginedProductPricer.java
@@ -7,9 +7,9 @@ package com.opengamma.strata.pricer.index;
 
 import com.opengamma.strata.market.sensitivity.PointSensitivities;
 import com.opengamma.strata.pricer.rate.RatesProvider;
-import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
 import com.opengamma.strata.product.index.IborFutureOption;
 import com.opengamma.strata.product.index.ResolvedIborFutureOption;
+import com.opengamma.strata.product.option.FutureOptionPremiumStyle;
 
 /**
  * Pricer for Ibor future option products with daily margin.
