+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/bond/BondDataSets.java
@@ -33,7 +33,7 @@ import com.opengamma.strata.product.bond.BondFutureOptionTrade;
 import com.opengamma.strata.product.bond.BondFutureTrade;
 import com.opengamma.strata.product.bond.FixedCouponBond;
 import com.opengamma.strata.product.bond.FixedCouponBondYieldConvention;
-import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
+import com.opengamma.strata.product.option.FutureOptionPremiumStyle;
 
 /**
  * Data sets of bond, bond future and bond future option.
