+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/index/IborFutureDummyData.java
@@ -17,12 +17,12 @@ import java.time.ZoneId;
 import com.opengamma.strata.basics.value.Rounding;
 import com.opengamma.strata.product.SecurityId;
 import com.opengamma.strata.product.TradeInfo;
-import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
 import com.opengamma.strata.product.common.PutCall;
 import com.opengamma.strata.product.index.IborFuture;
 import com.opengamma.strata.product.index.IborFutureOption;
 import com.opengamma.strata.product.index.IborFutureOptionTrade;
 import com.opengamma.strata.product.index.IborFutureTrade;
+import com.opengamma.strata.product.option.FutureOptionPremiumStyle;
 
 /**
  * Ibor future data.
