+++ modules/product/src/main/java/com/opengamma/strata/product/bond/BondFutureOptionSecurity.java
@@ -35,8 +35,8 @@ import com.opengamma.strata.product.Security;
 import com.opengamma.strata.product.SecurityId;
 import com.opengamma.strata.product.SecurityInfo;
 import com.opengamma.strata.product.TradeInfo;
-import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
 import com.opengamma.strata.product.common.PutCall;
+import com.opengamma.strata.product.option.FutureOptionPremiumStyle;
 
 /**
  * A security representing a futures contract, based on a basket of fixed coupon bonds.
