+++ modules/product/src/main/java/com/opengamma/strata/product/index/IborFutureOptionSecurity.java
@@ -37,8 +37,8 @@ import com.opengamma.strata.product.Security;
 import com.opengamma.strata.product.SecurityId;
 import com.opengamma.strata.product.SecurityInfo;
 import com.opengamma.strata.product.TradeInfo;
-import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
 import com.opengamma.strata.product.common.PutCall;
+import com.opengamma.strata.product.option.FutureOptionPremiumStyle;
 
 /**
  * A security representing a futures option contract, based on an Ibor index.
