+++ modules/product/src/main/java/com/opengamma/strata/product/option/package-info.java
@@ -1,0 +1,11 @@
+/**
+ * Copyright (C) 2016 - present by OpenGamma Inc. and the OpenGamma group of companies
+ * 
+ * Please see distribution for license.
+ */
+
+/**
+ * Entity objects describing common option concepts.
+ */
+package com.opengamma.strata.product.option;
+
