+++ modules/product/src/test/java/com/opengamma/strata/product/bond/BondFutureOptionSecurityTest.java
@@ -28,7 +28,7 @@ import com.opengamma.strata.product.SecurityId;
 import com.opengamma.strata.product.SecurityInfo;
 import com.opengamma.strata.product.SecurityPriceInfo;
 import com.opengamma.strata.product.TradeInfo;
-import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
+import com.opengamma.strata.product.option.FutureOptionPremiumStyle;
 
 /**
  * Test {@link BondFutureOptionSecurity}.
