+++ modules/product/src/test/java/com/opengamma/strata/product/fxopt/ResolvedFxSingleBarrierOptionTest.java
@@ -25,6 +25,9 @@ import org.testng.annotations.Test;
 
 import com.opengamma.strata.basics.currency.CurrencyAmount;
 import com.opengamma.strata.product.fx.ResolvedFxSingle;
+import com.opengamma.strata.product.option.BarrierType;
+import com.opengamma.strata.product.option.KnockType;
+import com.opengamma.strata.product.option.SimpleConstantContinuousBarrier;
 
 /**
  * Test {@link ResolvedFxSingleBarrierOption}.
