+++ modules/product/src/test/java/com/opengamma/strata/product/fxopt/ResolvedFxSingleBarrierOptionTradeTest.java
@@ -24,6 +24,9 @@ import com.opengamma.strata.basics.currency.CurrencyAmount;
 import com.opengamma.strata.basics.currency.Payment;
 import com.opengamma.strata.product.TradeInfo;
 import com.opengamma.strata.product.fx.ResolvedFxSingle;
+import com.opengamma.strata.product.option.BarrierType;
+import com.opengamma.strata.product.option.KnockType;
+import com.opengamma.strata.product.option.SimpleConstantContinuousBarrier;
 
 /**
  * Test {@link ResolvedFxSingleBarrierOptionTrade}.
