+++ modules/product/src/test/java/com/opengamma/strata/product/option/BarrierTypeTest.java
@@ -3,7 +3,7 @@
  * 
  * Please see distribution for license.
  */
-package com.opengamma.strata.product.fxopt;
+package com.opengamma.strata.product.option;
 
 import static com.opengamma.strata.collect.TestHelper.assertJodaConvert;
 import static com.opengamma.strata.collect.TestHelper.assertSerialization;
@@ -15,55 +15,55 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 /**
- * Test {@link KnockType}.
+ * Test {@link BarrierType}.
  */
 @Test
-public class KnockTypeTest {
+public class BarrierTypeTest {
 
   //-------------------------------------------------------------------------
-  public void test_isKnockIn() {
-    assertEquals(KnockType.KNOCK_IN.isKnockIn(), true);
-    assertEquals(KnockType.KNOCK_OUT.isKnockIn(), false);
+  public void test_isDown() {
+    assertEquals(BarrierType.UP.isDown(), false);
+    assertEquals(BarrierType.DOWN.isDown(), true);
   }
 
   //-------------------------------------------------------------------------
   @DataProvider(name = "name")
   static Object[][] data_name() {
     return new Object[][] {
-        {KnockType.KNOCK_IN, "KnockIn"},
-        {KnockType.KNOCK_OUT, "KnockOut"},
+        {BarrierType.UP, "Up"},
+        {BarrierType.DOWN, "Down"},
     };
   }
 
   @Test(dataProvider = "name")
-  public void test_toString(KnockType type, String name) {
+  public void test_toString(BarrierType type, String name) {
     assertEquals(type.toString(), name);
   }
 
   @Test(dataProvider = "name")
-  public void test_of_lookup(KnockType type, String name) {
-    assertEquals(KnockType.of(name), type);
+  public void test_of_lookup(BarrierType type, String name) {
+    assertEquals(BarrierType.of(name), type);
   }
 
   public void test_of_lookup_notFound() {
-    assertThrowsIllegalArg(() -> KnockType.of("Rubbish"));
+    assertThrowsIllegalArg(() -> BarrierType.of("Rubbish"));
   }
 
   public void test_of_lookup_null() {
-    assertThrowsIllegalArg(() -> KnockType.of(null));
+    assertThrowsIllegalArg(() -> BarrierType.of(null));
   }
 
   //-------------------------------------------------------------------------
   public void coverage() {
-    coverEnum(KnockType.class);
+    coverEnum(BarrierType.class);
   }
 
   public void test_serialization() {
-    assertSerialization(KnockType.KNOCK_OUT);
+    assertSerialization(BarrierType.DOWN);
   }
 
   public void test_jodaConvert() {
-    assertJodaConvert(KnockType.class, KnockType.KNOCK_IN);
+    assertJodaConvert(BarrierType.class, BarrierType.UP);
   }
 
 }
